backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
pydantic-settings==2.0.3
email-validator==2.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
alembic==1.12.0
bcrypt==4.0.1
psycopg2-binary==2.9.7
pytest==7.4.2
httpx==0.24.1